% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_bout_summary.R
\name{fm_bout_summary}
\alias{fm_bout_summary}
\title{Calculate mean true speed and flight duration for each bout}
\usage{
fm_bout_summary(df, ch_data, fm_total_duration)
}
\arguments{
\item{df}{Data frame to be modified. Use the data frame created from bouts function}

\item{ch_data}{The data frame for the individual chamber within the flight mill}

\item{fm_total_duration}{The total duration that the flight mill ran for. This is used to calculate the rest duration by taking away from the flight duration that is calculated}
}
\value{
A data frame with the bout flight parameters for the specified flight mill chamber
}
\description{
Summarises the mean true flight speed and the flight duration for each bout for the individual chamber
}
