% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_bouts.R
\name{fm_bouts}
\alias{fm_bouts}
\title{Summarising number of bouts in a flight run}
\usage{
fm_bouts(df, min_acceptable_duration)
}
\arguments{
\item{df}{data frame input}

\item{min_acceptable_duration}{The minimum acceptable duration of a stop to be considered a bout = FALSE. Duration of flight must be more than 2.5 seconds of sustained flight}
}
\value{
A data frame with flight bout parameters calculated
}
\description{
The number of bouts is calculated for each flight mill.
}
\note{
TO DO modify this to loop through several of the flight mills
}
