% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_series_plot.R
\name{fm_ch_plot}
\alias{fm_ch_plot}
\title{Time series plots}
\usage{
fm_ch_plot(df, title)
}
\arguments{
\item{df}{data frame input}

\item{title}{Name of the title of the plot}
}
\value{
A plot of the time series data for a single flight mill chamber
}
\description{
Plots for the time series data for a chamber in the flight mill
}
\note{
TO DO modify this to loop through several of the flight mill chambers
}
