% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_drop_rows.R
\name{fm_drop_nrows}
\alias{fm_drop_nrows}
\title{Drop the first n number of rows}
\usage{
fm_drop_nrows(df, N)
}
\arguments{
\item{df}{the dataframe required}

\item{N}{number of rows that need to be removed from the start of the file}
}
\value{
Output to a data frame with the first n number of rows removed
}
\description{
Drop the first set of rows from data frame. drops the fist n number of rows of the data as it is the first few rows that will be incorrect values as the time of the first two values will be part way through the hole in the disc.
}
