% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_duration_rest.R
\name{fm_duration_rest}
\alias{fm_duration_rest}
\title{Calculate the duration and rest for bouts}
\usage{
fm_duration_rest(df, bout)
}
\arguments{
\item{df}{Data frame to be modified. Use the data frame created from bouts function}

\item{bout}{The bout vector created from the bouts function.}
}
\value{
A data frame with the bout flight parameters for the specified flight mill chamber
}
\description{
Calculates the rest and duration for each bout in a flight mill chamber
}
