% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_filter_spikes.R
\name{fm_filter_spikes}
\alias{fm_filter_spikes}
\title{Identify anomalies in the flight mill data set}
\usage{
fm_filter_spikes(x, madThreshold = 4)
}
\arguments{
\item{x}{Data frame}

\item{madThreshold}{Computes the median absolute deviation using a numeric vector threshold}
}
\value{
A data frame with summary data and another data frame with the complete time series data
}
\description{
Identifies anomalies in the flight mill data set by using the median absolute deviation of the flight speed and creates a column that is either a 1: it is a spike and 0: is not a spike
}
