% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_mean_speed.R
\name{fm_mean_speed}
\alias{fm_mean_speed}
\alias{fm_speed}
\title{Mean speed of the flight mill chamber run}
\usage{
fm_speed(df, bout)
}
\arguments{
\item{df}{Data frame to be modified}

\item{bout}{Uses the vector bout that was created in the bouts function}
}
\value{
A data frame with flight parameters calculated for each bout.
}
\description{
Calculates the mean bout speed for each bout
}
