% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_series_plot.R
\name{fm_series_plot}
\alias{fm_series_plot}
\title{Time series plots}
\usage{
fm_series_plot(df, title)
}
\arguments{
\item{df}{data frame input}

\item{title}{Name of the title of the plot}
}
\value{
A plot of the time series data for a single flightmill
}
\description{
Plots for the time series data for each flightmill
}
\note{
TO DO modify this to loop through several of the flightmills
}
