% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_short_summary.R
\name{fm_short_summary}
\alias{fm_short_summary}
\title{Short flight mill Summary}
\usage{
fm_short_summary(df, bout_summary_data, duration_rest_data)
}
\arguments{
\item{df}{Data frame to be modified. Use the data frame created from bouts and rest_duration functions.}

\item{bout_summary_data}{Data frame of the bouts function.}

\item{duration_rest_data}{Vector created from the rest_duration function.}
}
\value{
A data frame with the bout flight parameters for the specified flight mill.
}
\description{
Summarises the flight mill data.
}
