% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_split_combine.R
\name{fm_split_combine}
\alias{fm_split_combine}
\title{Flight parameters in each mill and combine as two datasets}
\usage{
fm_split_combine(fm_name, minimum_bout, fm_duration)
}
\arguments{
\item{fm_name}{the flightmill name of the mill}

\item{minimum_bout}{The allowable time considered to be a bout for the species}

\item{fm_duration}{Length of time that the flight mill ran for}

\item{df}{Data frame to be used for calculating the flight parameters.}
}
\value{
A data frame with the flight parameters for the specified flight mill and another data frame with the time series data
}
\description{
Calculate all the flight parameters separately and then combine them as two complete datasets, a summary and a time series dataset. This combines the functions of
fm_bouts, fm_dr, fm_speed, fm_summary
}
