% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_summary.R
\name{fm_summary}
\alias{fm_summary}
\title{Summary of the flight mill data}
\usage{
fm_summary(bout_data, ch_data, ch_dr, ch_mean_speed, fm_total_duration)
}
\arguments{
\item{df}{Data frame to be modified. Use the data frame created from bouts and rest_dration functions.}

\item{bout_summary_data}{Data frame of the bouts function.}

\item{bout}{Vector created from the bouts function.}

\item{mean_speed_bout}{Vector created from the fm_speed data frame}
}
\value{
A data frame with the bout flight parameters for the specified flight mill chamber
}
\description{
Summarises the chamber data
}
