% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_summary.R
\name{fm_summary}
\alias{fm_summary}
\title{Summary of the flight mill data}
\usage{
fm_summary(bout_data, fm_data, fm_dr, fm_mean_speed, fm_total_duration)
}
\arguments{
\item{bout_data}{Data frame to be modified. Use the data frame created from bouts and rest_duration functions.}

\item{fm_data}{Flight mill data to be incorporated.}

\item{fm_dr}{Duration data to be included in the summary.}

\item{fm_mean_speed}{Vector created from the fm_speed data frame.}

\item{fm_total_duration}{The duration data to be included in the summary.}
}
\value{
A data frame with the bout flight parameters for the specified flight mill
}
\description{
Summarises the mill data
}
